sfendian: Fix possible arithmetic overflows

+++ src/sfendian.h
@@ -210,16 +210,16 @@ static inline int64_t
 psf_get_be64 (const uint8_t *ptr, int offset)
 {	int64_t value ;
 
-	value = ((uint32_t) ptr [offset]) << 24 ;
-	value += ptr [offset + 1] << 16 ;
-	value += ptr [offset + 2] << 8 ;
+	value = (int64_t) ((uint64_t) ptr [offset] << 24) ;
+	value += (int64_t) ((uint64_t) ptr [offset + 1] << 16) ;
+	value += (int64_t) ((uint64_t) ptr [offset + 2] << 8) ;
 	value += ptr [offset + 3] ;
 
 	value = (int64_t) (((uint64_t) value) << 32) ;
 
-	value += ((uint32_t) ptr [offset + 4]) << 24 ;
-	value += ptr [offset + 5] << 16 ;
-	value += ptr [offset + 6] << 8 ;
+	value += (int64_t) ((uint64_t) ptr [offset + 4] << 24) ;
+	value += (int64_t) ((uint64_t) ptr [offset + 5] << 16) ;
+	value += (int64_t) ((uint64_t) ptr [offset + 6] << 8) ;
 	value += ptr [offset + 7] ;
 	return value ;
 } /* psf_get_be64 */
@@ -228,16 +228,16 @@ static inline int64_t
 psf_get_le64 (const uint8_t *ptr, int offset)
 {	int64_t value ;
 
-	value = ((uint32_t) ptr [offset + 7]) << 24 ;
-	value += ptr [offset + 6] << 16 ;
-	value += ptr [offset + 5] << 8 ;
+	value = (int64_t) ((uint64_t) ptr [offset + 7] << 24) ;
+	value += (int64_t) ((uint64_t) ptr [offset + 6] << 16) ;
+	value += (int64_t) ((uint64_t) ptr [offset + 5] << 8) ;
 	value += ptr [offset + 4] ;
 
 	value = (int64_t) (((uint64_t) value) << 32) ;
 
-	value += ((uint32_t) ptr [offset + 3]) << 24 ;
-	value += ptr [offset + 2] << 16 ;
-	value += ptr [offset + 1] << 8 ;
+	value += (int64_t) ((uint64_t) ptr [offset + 3] << 24) ;
+	value += (int64_t) ((uint64_t) ptr [offset + 2] << 16) ;
+	value += (int64_t) ((uint64_t) ptr [offset + 1] << 8) ;
 	value += ptr [offset] ;
 	return value ;
 } /* psf_get_le64 */
